Fix VR mode lifecycle: The plugin now enters vr mode when initialize is invoked. This enables usage of the config GDScript classes right after the initialize invocation.
