+++ README.md
@@ -87,13 +87,13 @@ can be found in [src/config/](src/config/).
 
 Here is an example of how to use these classes inside GDScript:
 ```
-onready var ovr_init_config = preload("res://addons/godot_ovrmobile/OvrInitConfig.gdns").new()
-onready var ovr_runtime_config = preload("res://addons/godot_ovrmobile/OvrRuntimeConfig.gdns").new()
+onready var ovrTrackingTransform = preload("res://addons/godot_ovrmobile/OvrTrackingTransform.gdns").new()
+onready var ovrGuardianSystem = preload("res://addons/godot_ovrmobile/OvrGuardianSystem.gdns").new()
 
 func _process(delta):
-	print("GetTrackingSpace: " , ovr_runtime_config.get_tracking_space())
+	print("GetTrackingSpace: " , ovrTrackingTransform.get_tracking_space())
 
-	print("GetBoundaryVisible: ", ovr_runtime_config.get_boundary_visible())
-	print("GetBoundaryOrientedBoundingBox: ", ovr_runtime_config.get_boundary_oriented_bounding_box())
+	print("GetBoundaryVisible: ", ovrGuardianSystem.get_boundary_visible())
+	print("GetBoundaryOrientedBoundingBox: ", ovrGuardianSystem.get_boundary_oriented_bounding_box())
 ```
