+++ demo/Main.gd
@@ -1,7 +1,7 @@
 extends Spatial
 
 onready var ovr_init_config = preload("res://addons/godot_ovrmobile/OvrInitConfig.gdns").new()
-onready var ovr_runtime_config = preload("res://addons/godot_ovrmobile/OvrRuntimeConfig.gdns").new()
+onready var ovr_performance = preload("res://addons/godot_ovrmobile/OvrPerformance.gdns").new()
 
 func _ready():
 	# Get the count of available interfaces.
@@ -23,15 +23,10 @@ func _ready():
 		else:
 			print("Failed to enable OVRMobile")
 
-var first_time = true
 var perform_runtime_config = false
 
 func _process(delta):
-	if first_time:
-		print("First time processing.")
-		first_time = false
-	else:
-		if !perform_runtime_config:
-			ovr_runtime_config.set_clock_levels(1, 1)
-			ovr_runtime_config.set_extra_latency_mode(0)
-			perform_runtime_config = true
+	if !perform_runtime_config:
+		ovr_performance.set_clock_levels(1, 1)
+		ovr_performance.set_extra_latency_mode(1)
+		perform_runtime_config = true
