+++ demo/addons/godot_ovrmobile/OvrPerformance.gdns
@@ -3,6 +3,6 @@
 [ext_resource path="res://addons/godot_ovrmobile/godot_ovrmobile.gdnlib" type="GDNativeLibrary" id=1]
 
 [resource]
-resource_name = "OvrRuntimeConfig"
-class_name = "OvrRuntimeConfig"
+resource_name = "OvrPerformance"
+class_name = "OvrPerformance"
 library = ExtResource( 1 )
